requirements.txt
monai==1.0.1
manopth
#pyrender==0.1.33
smplx==0.1.28
hmr
scikit-learn
scikit-image
jpeg4py
matplotlib
pandas
configargparse
flatten_dict
chumpy
pytorch3d
bitsandbytes
opendr
numpy==1.23.1
yacs
networkx==2.5
charset-normalizer==3.1.0